Added different pages for multi-index support
