+++ src/Home.vue
@@ -1,0 +1,16 @@
+<template>
+    <div>It worked!</div>
+    <div>
+        <router-link to="/index1">Index 1</router-link>
+    </div>
+    <p>Example index with custom filter ordering, custom ordering of filter entries, and tabbed facets</p>
+    <div>
+        <router-link to="/index2">Index 2</router-link>
+    </div>
+    <p>Example index with a custom default sort ordering</p>
+</template>
+  
+export default {
+    name: 'Home'
+  };
+
