+++ src/main.ts
@@ -1,4 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import router from './router'; // Import the router
 
-createApp(App).mount('#app')
+createApp(App)
+  .use(router) // Use the router instance in your Vue app
+  .mount('#app');
